eliminar atributo

+++ backend/src/main/java/unpsjb/labprog/backend/presenter/AtributoPresenter.java
@@ -2,13 +2,19 @@ package unpsjb.labprog.backend.presenter;
 
 import unpsjb.labprog.backend.Response;
 import unpsjb.labprog.backend.business.AtributoService;
+
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+
 import unpsjb.labprog.backend.model.Atributo;
+
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -24,14 +30,15 @@ public class AtributoPresenter {
   AtributoService service;
 
   @GetMapping
-  public ResponseEntity<Object> findAll(@RequestParam(defaultValue = "0") int page,
+  public ResponseEntity<Object> findAll(@RequestParam(value = "filtered", required = false) boolean filtered,
+      @RequestParam(defaultValue = "0") int page,
       @RequestParam(defaultValue = "10") int size) {
-    return Response.ok(service.findByPage(service.findAll(), page, size));
+    return Response.ok(service.findByPage(service.findAllCategorias(filtered), page, size));
   }
 
   @RequestMapping(value = "/search", method = RequestMethod.GET)
-  public ResponseEntity<Object> search() {
-    return Response.ok(service.findAll());
+  public ResponseEntity<Object> search(@RequestParam(value = "filtered", required = false) boolean filtered) {
+    return Response.ok(service.findAllCategorias(filtered));
   }
 
   @PostMapping
@@ -42,6 +49,15 @@ public class AtributoPresenter {
         "Agenda creada correctamente");
   }
 
+  @PutMapping(value = "/{id}")
+  public ResponseEntity<Object> update(@PathVariable("id") long id) {
+    Atributo loteOrNull = service.findById(id);
+    loteOrNull.setDeleted(false);
+    return Response.ok(
+        service.add(loteOrNull),
+        "Agenda creada correctamente");
+  }
+
   @RequestMapping(value = "/id/{id}", method = RequestMethod.GET)
   public ResponseEntity<Object> findById(@PathVariable("id") int id) {
     Atributo loteOrNull = service.findById(id);
@@ -49,8 +65,15 @@ public class AtributoPresenter {
   }
 
   @DeleteMapping(value = "/delete/{id}")
-  public void delete(@PathVariable("id") Long id) {
+  public ResponseEntity<Object> delete(@PathVariable("id") long id) {
+    List<Atributo> categorias = service.findTipoDeProcesosByAtributo(id);
+    if (categorias.size() > 0) {
+      return Response.badRequest("");
+    }
     service.delete(id);
+    return Response.ok(
+        null,
+        "Categoria eliminada correctamente");
   }
 
 }
